src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.7.2
    # via django
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
dj-database-url==2.1.0
    # via -r src/requirements/requirements.in
django==4.1.13
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
idna==3.6
    # via requests
pillow==10.2.0
    # via -r src/requirements/requirements.in
psycopg2==2.9.9
    # via -r src/requirements/requirements.in
python-decouple==3.8
    # via -r src/requirements/requirements.in
requests==2.31.0
    # via
    #   -r src/requirements/requirements.in
    #   stripe
sqlparse==0.4.4
    # via django
stripe==8.7.0
    # via -r src/requirements/requirements.in
typing-extensions==4.10.0
    # via
    #   dj-database-url
    #   stripe
tzdata==2024.1
    # via django
urllib3==2.2.1
    # via requests
